Module higher::apply

Structs

ApplyFn

An ApplyFn is a function from A to B wrapped in something Rust’s type system can more easily digest. Arguments for Apply::apply() are required to be of this type rather than an arbitrary type matching Fn(A) -> B.

Traits

Apply

Apply takes an F<Fn(A) -> B> (or, rather, an F<ApplyFn<'a,A, B>> specifically) and applies it to an F<A> to produce an F<B>.

Functions

ap

apply_first

apply_second

f

lift2