Function higher::apply::lift2

pub fn lift2<'a, A, B: 'a, C: 'a, F, M>(
    f: &'a F,
    a: M,
    b: M::Target<B>
) -> M::Target<C>where
    A: Clone + 'a,
    F: Fn(A, B) -> C + 'a,
    M: Apply<'a, A>,
    M::Target<B>: Apply<'a, B, Target<C> = M::Target<C>> + Apply<'a, B, Target<ApplyFn<'a, B, C>> = M::Target<ApplyFn<'a, B, C>>>,
    M::Target<C>: Apply<'a, C, Target<B> = M::Target<B>> + Apply<'a, C, Target<ApplyFn<'a, B, C>> = M::Target<ApplyFn<'a, B, C>>>,

lift2 lifts a function of two arguments into an applicative.

“Lifting into” in this case means it converts a function which takes two values of types A and B and returns another value of type C into a function which takes Apply<A> and Apply<B> and returns an Apply<C> by applying the function to the values inside the applicatives.

This generic implmentation requires that the type A implements Clone, as each value of A may need to be passed to the function multiple times, depending on the structure of the applicative.