Trait higher::apply::Apply

pub trait Apply<'a, A>where
    A: 'a,{
    type Target<T>
    where
        T: 'a,
        A: 'a;

    fn apply<B>(self, f: Self::Target<ApplyFn<'a, A, B>>) -> Self::Target<B>
    where
        B: 'a;
}

Apply takes an F<Fn(A) -> B> (or, rather, an F<ApplyFn<'a,A, B>> specifically) and applies it to an F<A> to produce an F<B>.

Required Associated Types

type Target<T>
where
T: 'a,
A: 'a

Required Methods

fn apply<B>(self, f: Self::Target<ApplyFn<'a, A, B>>) -> Self::Target<B>where
B: 'a,

Apply an F of functions from A to B to an F of A, producing an F of B.