Trait higher::bifunctor::Bifunctor

pub trait Bifunctor<'a, A: 'a, B: 'a> {
    type Target<T: 'a, U: 'a>: Bifunctor<'a, T, U, Target<A, B> = Self>;

    fn bimap<C: 'a, D: 'a, L, R>(self, left: L, right: R) -> Self::Target<C, D>
    where
        L: Fn(A) -> C + 'a,
        R: Fn(B) -> D + 'a;

    fn lmap<C: 'a, L>(self, left: L) -> Self::Target<C, B>
    where
        Self: Sized,
        L: Fn(A) -> C + 'a,
    { ... }
    fn rmap<D: 'a, R>(self, right: R) -> Self::Target<A, D>
    where
        Self: Sized,
        R: Fn(B) -> D + 'a,
    { ... }
}

A Bifunctor lets you change the types of a generic type with two type parameters.

A Bifunctor works just like a Functor, but for types with two type parameters. It will convert a F<_, _>: Bifunctor from F<A, B> to F<C, D> using two functions, one Fn(A) -> C and the other Fn(B) -> D.

Required Associated Types

type Target<T: 'a, U: 'a>: Bifunctor<'a, T, U, Target<A, B> = Self>

Required Methods

fn bimap<C: 'a, D: 'a, L, R>(self, left: L, right: R) -> Self::Target<C, D>where
L: Fn(A) -> C + 'a,
R: Fn(B) -> D + 'a,

Map a Bifunctor<A, B> to a Bifunctor<C, D> using a function from A to C and a function from B to D.

Provided Methods

fn lmap<C: 'a, L>(self, left: L) -> Self::Target<C, B>where
Self: Sized,
L: Fn(A) -> C + 'a,

Map only the left hand side of the bifunctor from A to C.

fn rmap<D: 'a, R>(self, right: R) -> Self::Target<A, D>where
Self: Sized,
R: Fn(B) -> D + 'a,

Map only the right hand side of the bifunctor from B to D.