Trait higher::error::ApplicativeError

pub trait ApplicativeError<'a, A: 'a, E: 'a>: Functor<'a, A> + Pure<A> {
    fn throw_error(error: E) -> Self;
    fn handle_error_with<F: 'a>(self, f: F) -> Self
    where
        F: Fn(E) -> Self;

    fn handle_error<F: 'a>(self, f: F) -> Self
    where
        Self: Sized,
        F: Fn(E) -> A,
    { ... }
    fn attempt(self) -> Self::Target<Result<A, E>>
    where
        Self: Sized,
        Self::Target<Result<A, E>>: ApplicativeError<'a, A, E> + Pure<Result<A, E>>,
    { ... }
    fn recover_with<F>(self, recover: F) -> Self
    where
        Self: Sized,
        F: Fn(E) -> Result<Self, E> + 'a,
    { ... }
    fn recover<F>(self, recover: F) -> Self
    where
        Self: Sized,
        F: Fn(E) -> Result<A, E> + 'a,
    { ... }
    fn adapt_error<F>(self, adapt: F) -> Self
    where
        Self: Sized,
        F: Fn(E) -> E + 'a,
    { ... }
    fn redeem<B: 'a, FE: 'a, FA: 'a>(
        self,
        recover: FE,
        map: FA
    ) -> Self::Target<B>
    where
        Self: Sized,
        FE: Fn(E) -> B,
        FA: Fn(A) -> B,
        Self::Target<B>: ApplicativeError<'a, B, E>,
    { ... }
    fn from_option<F>(option: Option<A>, if_error: F) -> Self
    where
        Self: Sized,
        F: FnOnce() -> E,
    { ... }
    fn from_result(result: Result<A, E>) -> Self
    where
        Self: Sized,
    { ... }
}

Required Methods

fn throw_error(error: E) -> Self

Throw an error.

This constructs an error case for the ApplicativeError.

fn handle_error_with<F: 'a>(self, f: F) -> Selfwhere
F: Fn(E) -> Self,

Handle an error.

If the ApplicativeError contains an error, handle it using the provided function, returning a new ApplicativeError.

If there’s no error, do nothing.

Provided Methods

fn handle_error<F: 'a>(self, f: F) -> Selfwhere
Self: Sized,
F: Fn(E) -> A,

Handle an error.

If the ApplicativeError contains an error, handle it using the provided function, returning a success value which will be wrapped in a new ApplicativeError.

If there’s no error, do nothing.

fn attempt(self) -> Self::Target<Result<A, E>>where
Self: Sized,
Self::Target<Result<A, E>>: ApplicativeError<'a, A, E> + Pure<Result<A, E>>,

fn recover_with<F>(self, recover: F) -> Selfwhere
Self: Sized,
F: Fn(E) -> Result<Self, E> + 'a,

fn recover<F>(self, recover: F) -> Selfwhere
Self: Sized,
F: Fn(E) -> Result<A, E> + 'a,

fn adapt_error<F>(self, adapt: F) -> Selfwhere
Self: Sized,
F: Fn(E) -> E + 'a,

fn redeem<B: 'a, FE: 'a, FA: 'a>(self, recover: FE, map: FA) -> Self::Target<B>where
Self: Sized,
FE: Fn(E) -> B,
FA: Fn(A) -> B,
Self::Target<B>: ApplicativeError<'a, B, E>,

fn from_option<F>(option: Option<A>, if_error: F) -> Selfwhere
Self: Sized,
F: FnOnce() -> E,

fn from_result(result: Result<A, E>) -> Selfwhere
Self: Sized,