Trait higher::functor::FunctorRef

pub trait FunctorRef<'a, A: 'a>: Functor<'a, A> {
    fn fmap_ref<B: 'a, F: 'a>(&self, f: F) -> Self::Target<B>
    where
        F: Fn(&A) -> B;

    fn fclone(&self) -> Self
    where
        Self: Sized + FunctorRef<'a, A, Target<A> = Self>,
        A: Clone,
    { ... }
}

Required Methods

fn fmap_ref<B: 'a, F: 'a>(&self, f: F) -> Self::Target<B>where
F: Fn(&A) -> B,

Map a functor of A to a functor of B using a function from &A to B.

Provided Methods

fn fclone(&self) -> Selfwhere
Self: Sized + FunctorRef<'a, A, Target<A> = Self>,
A: Clone,

Given a type A implementing Clone, create a new identical FunctorRef<A> by cloning the values inside self.