Trait higher::FunctorRef

pub trait FunctorRef<'a, A: 'a>: Functor<'a, A> {
    fn fmap_ref<B: 'a, F>(&self, f: F) -> Self::Target<B>
    where
        F: Fn(&A) -> B + 'a;

    fn fclone(&self) -> Self
    where
        Self: Sized + FunctorRef<'a, A, Target<A> = Self>,
        A: Clone,
    { ... }
}

FunctorRef is an extension to Functor which provides a non-destructive fmap passing references to the mapping function.

This trait is separate from Functor because it can only be implemented for types which can be reconstructed using only references and the function mapping &A to B. For instance, it can’t be implemented for Result<A, E> because in the Err case, we can’t map to another Err without ownership of the E.

Required Methods

fn fmap_ref<B: 'a, F>(&self, f: F) -> Self::Target<B>where
F: Fn(&A) -> B + 'a,

Map a functor of A to a functor of B using a function from &A to B.

Provided Methods

fn fclone(&self) -> Selfwhere
Self: Sized + FunctorRef<'a, A, Target<A> = Self>,
A: Clone,

Given a type A implementing Clone, create a new identical FunctorRef<A> by cloning the values inside self.

This is mostly useful for data structures which don’t necessarily implement Clone. For those which do, you should reimplement this method simply as a call to Clone::clone() for performance.