Trait higher::Monoid

pub trait Monoid: Semigroup + Default { }

A Monoid consists of a Semigroup and an empty value (the Default trait) plus the following laws:

Associativity: (x + y) + z == x + (y + z)

Identity: 0 + a == a + 0 == a

If you’re wondering why this isn’t implemented for Rust’s primitive integers, it’s because they’re essentially two Monoids in a trenchcoat: both addition and multiplication fulfill the monoid laws. To distinguish between the two, you can wrap them in the Additive and Multiplicative newtypes. It’s generally easier to use Semiring as an abstraction for numbers for this reason, but the newtypes are there if you just need to use a number as a monoid.