Trait higher::Profunctor

pub trait Profunctor<'a, B: 'a, C: 'a> {
    type Target<T: 'a, U: 'a>;

    fn dimap<A: 'a, D: 'a, L, R>(self, left: L, right: R) -> Self::Target<A, D>
    where
        L: Fn(A) -> B + 'a,
        R: Fn(C) -> D + 'a;

    fn lcmap<A: 'a, L>(self, left: L) -> Self::Target<A, C>
    where
        Self: Sized,
        L: Fn(A) -> B + 'a,
    { ... }
    fn rmap<D: 'a, R>(self, right: R) -> Self::Target<B, D>
    where
        Self: Sized,
        R: Fn(C) -> D + 'a,
    { ... }
}

A Profunctor is just a Bifunctor that is contravariant over its first argument and covariant over its second argument. What’s the problem?

Required Associated Types

type Target<T: 'a, U: 'a>

Required Methods

fn dimap<A: 'a, D: 'a, L, R>(self, left: L, right: R) -> Self::Target<A, D>where
L: Fn(A) -> B + 'a,
R: Fn(C) -> D + 'a,

Map a function over both arguments of the profunctor.

Provided Methods

fn lcmap<A: 'a, L>(self, left: L) -> Self::Target<A, C>where
Self: Sized,
L: Fn(A) -> B + 'a,

Map a function over the contravariant first argument only.

fn rmap<D: 'a, R>(self, right: R) -> Self::Target<B, D>where
Self: Sized,
R: Fn(C) -> D + 'a,

Map a function over the covariant second argument only.