Trait higher::apply::Apply

pub trait Apply<'a, A>where
    A: 'a,{
    type Target<T>
    where
        T: 'a,
        A: 'a;

    fn apply<B>(self, f: Self::Target<ApplyFn<'a, A, B>>) -> Self::Target<B>
    where
        B: 'a;

    fn apply_first<B, MB, MF>(self, b: MB) -> Self
    where
        Self: Sized + Apply<'a, A, Target<B> = MB> + Functor<'a, A, Target<ApplyFn<'a, B, A>> = MF>,
        A: Clone,
        B: 'a,
        MB: Apply<'a, B, Target<ApplyFn<'a, B, A>> = MF> + Apply<'a, B, Target<A> = Self>,
        MF: Apply<'a, ApplyFn<'a, B, A>, Target<B> = MB>,
    { ... }
    fn apply_second<B, MB, MF>(self, b: MB) -> MB
    where
        Self: Sized + Apply<'a, A, Target<B> = MB> + Apply<'a, A, Target<ApplyFn<'a, A, B>> = MF>,
        B: Clone + 'a,
        MB: Apply<'a, B, Target<ApplyFn<'a, A, B>> = MF> + Functor<'a, B, Target<ApplyFn<'a, A, B>> = MF>,
        MF: Apply<'a, ApplyFn<'a, A, B>, Target<A> = Self>,
    { ... }
}

Apply takes an F<Fn(A) -> B> (or, rather, an F<ApplyFn<'a,A, B>> specifically) and applies it to an F<A> to produce an F<B>.

Required Associated Types

type Target<T>
where
T: 'a,
A: 'a

Required Methods

fn apply<B>(self, f: Self::Target<ApplyFn<'a, A, B>>) -> Self::Target<B>where
B: 'a,

Apply an F of functions from A to B to an F of A, producing an F of B.

Provided Methods

fn apply_first<B, MB, MF>(self, b: MB) -> Selfwhere
Self: Sized + Apply<'a, A, Target<B> = MB> + Functor<'a, A, Target<ApplyFn<'a, B, A>> = MF>,
A: Clone,
B: 'a,
MB: Apply<'a, B, Target<ApplyFn<'a, B, A>> = MF> + Apply<'a, B, Target<A> = Self>,
MF: Apply<'a, ApplyFn<'a, B, A>, Target<B> = MB>,

fn apply_second<B, MB, MF>(self, b: MB) -> MBwhere
Self: Sized + Apply<'a, A, Target<B> = MB> + Apply<'a, A, Target<ApplyFn<'a, A, B>> = MF>,
B: Clone + 'a,
MB: Apply<'a, B, Target<ApplyFn<'a, A, B>> = MF> + Functor<'a, B, Target<ApplyFn<'a, A, B>> = MF>,
MF: Apply<'a, ApplyFn<'a, A, B>, Target<A> = Self>,