Trait higher::apply::Apply

pub trait Apply<'a, A: 'a>: Functor<'a, A> {
    fn apply<B: 'a>(self, f: Self::Target<ApplyFn<'a, A, B>>) -> Self::Target<B>;

    fn apply_first<B: 'a>(self, b: Self::Target<B>) -> Self
    where
        Self: Sized,
        A: Clone,
        Self::Target<B>: Apply<'a, B, Target<ApplyFn<'a, B, A>> = Self::Target<ApplyFn<'a, B, A>>>,
    { ... }
    fn apply_second<B>(self, b: Self::Target<B>) -> Self::Target<B>
    where
        Self: Sized,
        B: Clone + 'a,
        Self::Target<B>: Apply<'a, B, Target<ApplyFn<'a, A, B>> = Self::Target<ApplyFn<'a, A, B>>>,
    { ... }
}

Apply takes an Apply<Fn(A) -> B> (or, rather, an Apply<ApplyFn<'a, A, B>> specifically) and applies it to an Apply<A> to produce an Apply<B>.

In simpler terms, it takes a value of type A inside some context (“some context” here being the type which implements Apply) and a function Fn(A) ->B inside a similar context and produces a similar context containing the result (B) of applying the function to the value. Some concrete examples of “some context” are Option, Vec and futures.

Required Methods

fn apply<B: 'a>(self, f: Self::Target<ApplyFn<'a, A, B>>) -> Self::Target<B>

Apply an F of functions from A to B to an F of A, producing an F of B.

Provided Methods

fn apply_first<B: 'a>(self, b: Self::Target<B>) -> Selfwhere
Self: Sized,
A: Clone,
Self::Target<B>: Apply<'a, B, Target<ApplyFn<'a, B, A>> = Self::Target<ApplyFn<'a, B, A>>>,

fn apply_second<B>(self, b: Self::Target<B>) -> Self::Target<B>where
Self: Sized,
B: Clone + 'a,
Self::Target<B>: Apply<'a, B, Target<ApplyFn<'a, A, B>> = Self::Target<ApplyFn<'a, A, B>>>,