Trait higher::functor::Functor

pub trait Functor<'a, A>where
    A: 'a,{
    type Target<T>
    where
        T: 'a;

    fn fmap<B, F>(self, f: F) -> Self::Target<B>
    where
        B: 'a,
        F: Fn(A) -> B + 'a;

    fn fconst<B>(self, b: B) -> Self::Target<B>
    where
        Self: Sized,
        B: Clone,
    { ... }
    fn void(self) -> Self::Target<()>
    where
        Self: Sized,
    { ... }
    fn f_into_iter(self) -> Box<dyn Iterator<Item = A>>
    where
        Self: Sized,
        A: 'static,
    { ... }
}

A Functor lets you change the type parameter of a generic type.

A Functor defines a method fmap on a type F<_>: Functor which converts an F<A> to F<B> using a function Fn(A) -> B applied to the As inside it.

You can also use this just to modify the values inside your container value without changing their type, if the mapping function returns a value of the same type. This is called an “endofunctor.”

Required Associated Types

type Target<T>
where
T: 'a

Required Methods

fn fmap<B, F>(self, f: F) -> Self::Target<B>where
B: 'a,
F: Fn(A) -> B + 'a,

Map a functor of A to a functor of B using a function from A to B.

Provided Methods

fn fconst<B>(self, b: B) -> Self::Target<B>where
Self: Sized,
B: Clone,

Map the functor to the provided constant value.

fn void(self) -> Self::Target<()>where
Self: Sized,

Map the functor to the unit value ().

fn f_into_iter(self) -> Box<dyn Iterator<Item = A>>where
Self: Sized,
A: 'static,

Turn the functor into an iterator.

let my_functor = vec![1, 2, 3];
let iter = my_functor.f_into_iter();
let my_vec: Vec<i32> = iter.collect();
assert_eq!(my_vec, vec![1, 2, 3]);